issue-#56 - removed ProxyHandlerProvider.java
