+++ src/main/java/com/networknt/proxy/LightProxyHandler.java
@@ -30,8 +30,7 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Arrays;
 import java.util.List;
-
-import static com.networknt.proxy.ProxyHandlerProvider.handlingConsumerWrapper;
+import java.util.function.Consumer;
 
 /**
  * This is a wrapper class for LightProxyHandler as it is implemented as final. This class implements
@@ -43,6 +42,7 @@ import static com.networknt.proxy.ProxyHandlerProvider.handlingConsumerWrapper;
 public class LightProxyHandler implements HttpHandler {
     static final String CONFIG_NAME = "proxy";
     static ProxyConfig config = (ProxyConfig) Config.getInstance().getJsonObjectConfig(CONFIG_NAME, ProxyConfig.class);
+    static final Logger logger = LoggerFactory.getLogger(LightProxyHandler.class);
 
     ProxyHandler proxyHandler;
 
@@ -78,5 +78,26 @@ public class LightProxyHandler implements HttpHandler {
     @Override
     public void handleRequest(HttpServerExchange httpServerExchange) throws Exception {
         proxyHandler.handleRequest(httpServerExchange);
+    }
+
+    @FunctionalInterface
+    public interface ThrowingConsumer<T, E extends Exception> {
+        void accept(T t) throws E;
+    }
+    static <T, E extends Exception> Consumer<T> handlingConsumerWrapper(
+            ThrowingConsumer<T, E> throwingConsumer, Class<E> exceptionClass) {
+
+        return i -> {
+            try {
+                throwingConsumer.accept(i);
+            } catch (Exception ex) {
+                try {
+                    E exCast = exceptionClass.cast(ex);
+                    logger.error("Exception occured :", ex);
+                } catch (ClassCastException ccEx) {
+                    throw new RuntimeException(ex);
+                }
+            }
+        };
     }
 }
